Add OrderId property to Order entity and update migration
Updated ApplicationDbContextModelSnapshot.cs to include a new
required property OrderId in the Order entity with column type
nvarchar(max). Added a new migration file 20241105224313_UpdateOrderTable.cs
to add the OrderId column to the Orders table in the database,
ensuring it is not nullable with a default value of an empty string.
Generated 20241105224313_UpdateOrderTable.Designer.cs to reflect
these changes in the model snapshot.