Add FileDownloadResult class to Models namespace
A new class `FileDownloadResult` has been added to the `ST10361554_CLDV6212_POE_Part_3_WebApp.Models` namespace. This class represents the result of a file download operation and includes three required properties: `Content` (a `MemoryStream` containing the file's content), `ContentType` (a `string` specifying the MIME type of the file), and `FileName` (a `string` specifying the name of the file to be downloaded).