Add comments to DbSet properties in ApplicationDbContext
Added comments to the DbSet properties in ApplicationDbContext.cs
to indicate the purpose of each DbSet. These comments specify that
the Users property represents the "User table", the Products
property represents the "Product table", and the Orders property
represents the "Order table". These changes are for documentation
purposes only and do not affect functionality.